fix: complete bridge migration to *_impl naming, fix build warnings, add Terminal UAPI paths
- Bridge naming: all host_* entrypoints → *_impl() (socket_impl, bind_impl, log_init_impl, etc.)
- Bridge headers: created include/ixland/{sock_bridge.h,log_bridge.h,platform_bridge.h} for Linux-owner visibility
- Deleted include/ixland/host_socket_types.h (host ABI types no longer leak into Linux-owner code)
- ABI boundary: struct bridge_msghdr/bridge_iovec translation happens entirely in internal/ios/fs/sock_bridge.c
- Fixed -Wshorten-64-to-32 warnings across 20+ files (explicit casts for int64_t→int, size_t→uint32_t, ssize_t→int)
- Fixed -Wstrict-prototypes warnings (K&R () → (void) in 10+ files)
- Fixed -Wformat warnings in kernel/aarch64/calls.c and emu/aarch64/cpu.c (int64_t with %lld)
- Fixed -Wunused-function warnings in emu/aarch64/cpu.c (__attribute__((unused)) on 4 trace functions)
- Fixed -Wconditional-uninitialized in emu/aarch64/cpu.c (uint64_t base = 0)
- Fixed -Wunreachable-code in tcti/aarch64/gen.c (removed dead code after switch)
- Added /include and  to IXLandTerminal HEADER_SEARCH_PATHS (fakefs.c needs runtime headers)
- Regenerated IXLand.xcodeproj from project.yml
- IXLandLinuxRuntime scheme: BUILD SUCCEEDED
- Vendored UAPI provenance verified: 863 headers with GPL-2.0 WITH Linux-syscall-note from make headers_install
- No host header leaks in Linux-owner code (no <netinet/>, <sys/socket.h>, <os/log.h>)
- No active host_* bridge naming violations (only internal variable names like host_ptr, host_tid remain)
- internal/ios/ quarantine intact

+++ Sources/IXLandLinuxRuntime/emu/aarch64/decode.c
@@ -559,7 +559,7 @@ int a64_decode_ldst(uint32_t insn, a64_instr_t *out)
             break;
         }
 
-        out->pair_offset = sign_extend(imm7, 7)
+        out->pair_offset = (int)sign_extend(imm7, 7)
                            << (out->is_vector ? (2 + out->size) : (2 + (op0 >> 1)));
         return 0;
     }
