+++ Sources/IXLandLinuxRuntime/fs/fake-rebuild.c
@@ -98,11 +98,11 @@ int fakefs_rebuild(struct fakefs_db *fs, int root_fd)
         // store all the information in the new database
         err = sqlite3_bind_int64(write_stat, 1, real_inode);
         CHECK_ERR();
-        err = sqlite3_bind_blob(write_stat, 2, stat_data, stat_data_size, SQLITE_TRANSIENT);
+        err = sqlite3_bind_blob(write_stat, 2, stat_data, (int)stat_data_size, SQLITE_TRANSIENT);
         CHECK_ERR();
         STEP(write_stat);
         RESET(write_stat);
-        err = sqlite3_bind_blob(write_path, 1, path, strlen(path), SQLITE_TRANSIENT);
+        err = sqlite3_bind_blob(write_path, 1, path, (int)strlen(path), SQLITE_TRANSIENT);
         CHECK_ERR();
         err = sqlite3_bind_int64(write_path, 2, real_inode);
         CHECK_ERR();
