+++ Sources/IXLandLinuxRuntime/fs/proc.c
@@ -1,11 +1,12 @@
-#include <string.h>
-#include <sys/stat.h>
+#import <IXLandLinuxRuntime/fs/path.h>
+#import <IXLandLinuxRuntime/fs/proc.h>
 #import <IXLandLinuxRuntime/kernel/calls.h>
 #import <IXLandLinuxRuntime/kernel/fs.h>
-#import <IXLandLinuxRuntime/fs/proc.h>
-#import <IXLandLinuxRuntime/fs/path.h>
+#include <string.h>
+#include <sys/stat.h>
 
-static int proc_lookup(const char *path, struct proc_entry *entry) {
+static int proc_lookup(const char *path, struct proc_entry *entry)
+{
     entry->meta = &proc_root;
     char component[MAX_NAME + 1];
     int err = 0;
@@ -16,7 +17,7 @@ static int proc_lookup(const char *path, struct proc_entry *entry) {
         }
 
         unsigned long index = 0;
-        struct proc_entry next_entry = {0};
+        struct proc_entry next_entry = { 0 };
         char entry_name[MAX_NAME];
         while (proc_dir_read(entry, &index, &next_entry)) {
             // tack on some dynamically generated attributes
@@ -26,7 +27,7 @@ static int proc_lookup(const char *path, struct proc_entry *entry) {
                 // this asserts that an entry has a unique parent
                 assert(next_entry.meta->parent == entry->meta);
             if (next_entry.meta->inode == 0)
-                next_entry.meta->inode = index + 1;
+                next_entry.meta->inode = (int)(index + 1);
 
             proc_entry_getname(&next_entry, entry_name);
             if (strcmp(entry_name, component) == 0)
@@ -46,8 +47,10 @@ found:
 
 extern const struct fd_ops procfs_fdops;
 
-static struct fd *proc_open(struct mount *UNUSED(mount), const char *path, int UNUSED(flags), int UNUSED(mode)) {
-    struct proc_entry entry = {0};
+static struct fd *proc_open(struct mount *UNUSED(mount), const char *path, int UNUSED(flags),
+                            int UNUSED(mode))
+{
+    struct proc_entry entry = { 0 };
     int err = proc_lookup(path, &entry);
     if (err < 0)
         return ERR_PTR(err);
@@ -57,7 +60,8 @@ static struct fd *proc_open(struct mount *UNUSED(mount), const char *path, int U
     return fd;
 }
 
-static int proc_getpath(struct fd *fd, char *buf) {
+static int proc_getpath(struct fd *fd, char *buf)
+{
     char *p = buf + MAX_PATH - 1;
     size_t n = 0;
     p[0] = '\0';
@@ -76,8 +80,9 @@ static int proc_getpath(struct fd *fd, char *buf) {
     return 0;
 }
 
-static int proc_stat(struct mount *UNUSED(mount), const char *path, struct statbuf *stat) {
-    struct proc_entry entry = {0};
+static int proc_stat(struct mount *UNUSED(mount), const char *path, struct statbuf *stat)
+{
+    struct proc_entry entry = { 0 };
     int err = proc_lookup(path, &entry);
     if (err < 0)
         return err;
@@ -86,11 +91,13 @@ static int proc_stat(struct mount *UNUSED(mount), const char *path, struct statb
     return ret;
 }
 
-static int proc_fstat(struct fd *fd, struct statbuf *stat) {
+static int proc_fstat(struct fd *fd, struct statbuf *stat)
+{
     return proc_entry_stat(&fd->proc.entry, stat);
 }
 
-static int proc_refresh_data(struct fd *fd) {
+static int proc_refresh_data(struct fd *fd)
+{
     mode_t_ mode = proc_entry_mode(&fd->proc.entry);
     if (S_ISDIR(mode))
         return _EISDIR;
@@ -108,7 +115,8 @@ static int proc_refresh_data(struct fd *fd) {
     return 0;
 }
 
-static off_t_ proc_seek(struct fd *fd, off_t_ off, int whence) {
+static off_t_ proc_seek(struct fd *fd, off_t_ off, int whence)
+{
     int err = proc_refresh_data(fd);
     if (err < 0)
         return err;
@@ -120,12 +128,13 @@ static off_t_ proc_seek(struct fd *fd, off_t_ off, int whence) {
     return fd->offset;
 }
 
-static ssize_t proc_pread(struct fd *fd, void *buf, size_t bufsize, off_t off) {
+static ssize_t proc_pread(struct fd *fd, void *buf, size_t bufsize, off_t off)
+{
     if (fd->proc.entry.meta->pread) {
-        struct proc_data data = {buf, bufsize, bufsize};
+        struct proc_data data = { buf, bufsize, bufsize };
         return fd->proc.entry.meta->pread(&fd->proc.entry, &data, off);
     }
-    
+
     int err = proc_refresh_data(fd);
     if (err < 0)
         return err;
@@ -134,7 +143,7 @@ static ssize_t proc_pread(struct fd *fd, void *buf, size_t bufsize, off_t off) {
     assert(data != NULL);
 
     size_t remaining = fd->proc.data.size - off;
-    if ((size_t) off > fd->proc.data.size)
+    if ((size_t)off > fd->proc.data.size)
         remaining = 0;
     size_t n = bufsize;
     if (n > remaining)
@@ -144,7 +153,8 @@ static ssize_t proc_pread(struct fd *fd, void *buf, size_t bufsize, off_t off) {
     return n;
 }
 
-static void proc_buf_write(struct proc_data *buf, const void *data, size_t size, size_t off) {
+static void proc_buf_write(struct proc_data *buf, const void *data, size_t size, size_t off)
+{
     assert(off <= buf->size);
     size_t tidemark = off + size;
     if (tidemark > buf->capacity) {
@@ -166,29 +176,31 @@ static void proc_buf_write(struct proc_data *buf, const void *data, size_t size,
     }
 }
 
-static ssize_t proc_pwrite(struct fd *fd, const void *buf, size_t bufsize, off_t off) {
+static ssize_t proc_pwrite(struct fd *fd, const void *buf, size_t bufsize, off_t off)
+{
     mode_t_ mode = proc_entry_mode(&fd->proc.entry);
     if (S_ISDIR(mode))
         return _EISDIR;
     assert(S_ISREG(mode));
-    
+
     if (fd->proc.entry.meta->pwrite) {
-        struct proc_data data = {(char *) buf, bufsize, bufsize};
+        struct proc_data data = { (char *)buf, bufsize, bufsize };
         return fd->proc.entry.meta->pwrite(&fd->proc.entry, &data, off);
     }
-    
+
     if (!fd->proc.entry.meta->update) {
         return _EPERM;
     }
-    
-    struct proc_data data = {(char *)buf, bufsize, bufsize};
+
+    struct proc_data data = { (char *)buf, bufsize, bufsize };
     fd->proc.entry.meta->update(&fd->proc.entry, &data);
-    
+
     return bufsize;
 }
 
-static int proc_readdir(struct fd *fd, struct dir_entry *entry) {
-    struct proc_entry proc_entry = {0};
+static int proc_readdir(struct fd *fd, struct dir_entry *entry)
+{
+    struct proc_entry proc_entry = { 0 };
     bool any_left = proc_dir_read(&fd->proc.entry, &fd->offset, &proc_entry);
     if (!any_left)
         return 0;
@@ -198,7 +210,8 @@ static int proc_readdir(struct fd *fd, struct dir_entry *entry) {
     return 1;
 }
 
-static int proc_close(struct fd *fd) {
+static int proc_close(struct fd *fd)
+{
     if (fd->proc.data.data != NULL)
         free(fd->proc.data.data);
     proc_entry_cleanup(&fd->proc.entry);
@@ -213,8 +226,10 @@ const struct fd_ops procfs_fdops = {
     .close = proc_close,
 };
 
-static ssize_t proc_readlink(struct mount *UNUSED(mount), const char *path, char *buf, size_t bufsize) {
-    struct proc_entry entry = {0};
+static ssize_t proc_readlink(struct mount *UNUSED(mount), const char *path, char *buf,
+                             size_t bufsize)
+{
+    struct proc_entry entry = { 0 };
     int err = proc_lookup(path, &entry);
     if (err < 0)
         return err;
@@ -234,8 +249,9 @@ static ssize_t proc_readlink(struct mount *UNUSED(mount), const char *path, char
     return bufsize;
 }
 
-static int proc_unlink(struct mount *UNUSED(mount), const char *path) {
-    struct proc_entry entry = {0};
+static int proc_unlink(struct mount *UNUSED(mount), const char *path)
+{
+    struct proc_entry entry = { 0 };
     int err = proc_lookup(path, &entry);
     if (err < 0)
         return err;
@@ -246,11 +262,13 @@ static int proc_unlink(struct mount *UNUSED(mount), const char *path) {
     return err;
 }
 
-void proc_buf_append(struct proc_data *buf, const void *data, size_t size) {
+void proc_buf_append(struct proc_data *buf, const void *data, size_t size)
+{
     proc_buf_write(buf, data, size, buf->size);
 }
 
-void proc_printf(struct proc_data *buf, const char *format, ...) {
+void proc_printf(struct proc_data *buf, const char *format, ...)
+{
     char data[4096];
     va_list args;
     va_start(args, format);
@@ -260,7 +278,8 @@ void proc_printf(struct proc_data *buf, const char *format, ...) {
 }
 
 const struct fs_ops procfs = {
-    .name = "proc", .magic = 0x9fa0,
+    .name = "proc",
+    .magic = 0x9fa0,
     .open = proc_open,
     .getpath = proc_getpath,
     .stat = proc_stat,
